Update the scene_steps.rb definitions to better accommodate the working of the "add scene" command. Add a replacement feature expressing how multi-part scene contents could be expressed in feature lists.
Make the "@game" objects in the steps into a local variable (method, actually) singleton to more reliably track changes in previous feature steps.
Add a TODO item tracking the need to refactor this later.

Change the name of the "to_scene" command to "run_scene"
Add preliminary support for the "run_scene" command.