Add initial, fulfilled tests for explicitly displaying characters on the stage.
Update existing tests to reflect the addtion of the character to the stage and the addition of the character's description.

Update the character output to include the character's descriptions almost across the board.  (Replaces the display of the character's image URL used in the description's place.)

Update the tutorial's CSS to hide the character summaries.